pyproject.toml
[tool.poetry]
name = "projeto-sistemas-distribuidos"
version = "0.1.0"
description = ""
authors = ["Otávio Leite <atoivat.o@gmail.com>"]
readme = "README.md"
packages = [{include = "projeto_sistemas_distribuidos"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
grpcio = "^1.53.0"
grpcio-tools = "^1.53.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.2"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.master]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "invalid-name",
    "too-few-public-methods"
]
